Changed ActivitiesService to ActivityService, its path from activities to activity, getAll to list, getOne to retrieve; Updated imports to new path and class; Added docstrings to UnitActivity related services, model, enum, pipe;
